+++ R/shiny-driver.R
@@ -90,6 +90,9 @@
 #'   \item{timeout}{Timeout for the condition, in milliseconds.}
 #'   \item{output}{Character vector, the name(s) of the Shiny output
 #'     widgets that should be updated.}
+#'   \item{allowInputNoBinding_}{When setting the value of an input, allow
+#'     it to set the value of an input even if that input does not have
+#'     an input binding.}
 #'   \item{...}{For \code{expectUpdate} these can be named arguments.
 #'     The argument names correspond to Shiny input widgets: each input
 #'     widget will be set to the specified value.}
@@ -310,9 +313,11 @@ ShinyDriver <- R6Class(
       sd_expectUpdate(self, private, output, ..., timeout = timeout,
                        iotype = match.arg(iotype)),
 
-    setInputs = function(..., wait_ = TRUE, values_ = TRUE, timeout_ = 3000) {
+    setInputs = function(..., wait_ = TRUE, values_ = TRUE, timeout_ = 3000,
+      allowInputNoBinding_ = FALSE, priority_ = c("input", "event")) {
       sd_setInputs(self, private, ..., wait_ = wait_, values_ = values_,
-                    timeout_ = timeout_)
+                   timeout_ = timeout_, allowInputNoBinding_ = allowInputNoBinding_,
+                   priority_ = priority_)
     },
 
     uploadFile = function(..., wait_ = TRUE, values_ = TRUE, timeout_ = 3000)