+++ inst/recorder/recorder.js
@@ -45,13 +45,15 @@ window.shinyRecorder = (function() {
         }
 
         // Check if value has changed from last time.
-        var valueJSON = JSON.stringify(event.value);
-        if (valueJSON === previousInputValues[event.name])
-            return;
-        previousInputValues[event.name] = valueJSON;
+        if (event.priority !== "event") {
+            var valueJSON = JSON.stringify(event.value);
+            if (valueJSON === previousInputValues[event.name])
+                return;
+            previousInputValues[event.name] = valueJSON;
+        }
 
         var hasBinding = !!event.binding;
-        sendInputEventToParent(event.inputType, event.name, event.value, hasBinding);
+        sendInputEventToParent(event.inputType, event.name, event.value, hasBinding, event.priority);
     });
 
     $(document).on("shiny:filedownload", function(event) {
@@ -74,6 +76,15 @@ window.shinyRecorder = (function() {
     $(document).on("shiny:updateinput", function(event) {
         var inputId = event.binding.getId(event.target);
         updatedInputs[inputId] = true;
+        // Schedule this updated input to be cleared at the end of this tick.
+        // This is useful in the case where an input is updated with an empty
+        // value -- for example, if a selectInput is updated with a number of
+        // selections and a value of character(0), then it will not be removed
+        // from the updatedInputs list via the other code paths. (Note that it
+        // is possible in principle for other functions to be scheduled to
+        // occur afterward, but on the same tick, but in practice this
+        // shouldn't occur.)
+        setTimeout(function() { delete updatedInputs[inputId]; }, 0);
     });
 
     // Ctrl-click or Cmd-click (Mac) to record an output value
@@ -88,6 +99,17 @@ window.shinyRecorder = (function() {
         sendOutputSnapshotToParent($el[0].id);
     });
 
+    // Trigger a snapshot on Ctrl-shift-S or Cmd-shift-S (Mac)
+    $(document).keydown(function(e) {
+        if (!(e.ctrlKey || e.metaKey))
+            return;
+        if (!e.shiftKey)
+            return;
+        if (e.which !== 83)
+            return;
+
+        sendSnapshotToParent();
+    });
 
     function debounce(f, delay) {
         var timer = null;
@@ -101,14 +123,15 @@ window.shinyRecorder = (function() {
         };
     }
 
-    function sendInputEventToParent(inputType, name, value, hasBinding) {
+    function sendInputEventToParent(inputType, name, value, hasBinding, priority) {
         parent.postMessage({
             token: shinyrecorder.token,
             inputEvent: {
                 inputType: inputType,
                 name: name,
                 value: value,
-                hasBinding: hasBinding
+                hasBinding: hasBinding,
+                priority: priority
              }
         }, "*");
     }
@@ -131,6 +154,13 @@ window.shinyRecorder = (function() {
     // output events will all happen in a single tick. Debouncing for one tick
     // will collapse them into a single call to sendOutputEventToParent().
     var sendOutputEventToParentDebounced = debounce(sendOutputEventToParent, 10);
+
+    function sendSnapshotToParent() {
+        parent.postMessage({
+            token: shinyrecorder.token,
+            snapshotKeypress: true
+        }, "*");
+    }
 
     function sendOutputSnapshotToParent(name) {
         parent.postMessage({